Create Clear-ClipboardContents.ps1
This script can be run as a scheduled task and run silently in the background (-WindowStyle  Hidden -NonInteractive). Trigger it on logon and startup.

The script will monitor the contents of your clipboard by checking its hash value.
If the clipboard contents have not changed in 2 minutes, it will flush the clipboard.
This is so that IF you ever had an attacker, they cannot grab any information stored in your clipboard, which can be a very nice little treason trove at times.